Fix for UNION SELECT queries which are enclosed in parentheses
In UNION SELECT queries where additional processing must be performed
on the united data, the initial parentheses which are required to
encapsulate the subqueries correctly cause Pods's query detection at
line 1919 to fail. This pull request fixes the issue and also allows
for additional whitespace characters before the SELECT keyword.