- adding the watch module
The watch module allows us to watch files and do auto-building of
items as they change.

So as I'm working on the Sass stuff that means I can run the `grunt
watch` command and then as I change Sass files Grunt will auto-build
things for me.

Of course that means I'll need to build the watch and sass tasks shortly
as well.