- our starting grunt build process
This commit has a validated build process (and watch commands) running
for our Sass.

Now you can runn `grunt watch` and that will give you a terminal task
that watches the sass files and builds them.

Right now it's only the one file, but next up is adding all the other
CSS files that we need.