Properly build related object table
If pods_register_related_object() is called late in the game then the objects aren't actually added to the array
because the optimizations added for 2.5 keep the tables from constantly being rebuilded.  This patch lets the items to
be added without doing a full rebuild.  Also, a force rebuild before would have loaded the transient and thrown the
data away.

Fixes #2737