New Link field with WP link modal
Adds a link field type with options to add a link with a title and
target.
Validation similar to website field.

UI has the option to open the WordPress link modal to search for
existing content and autofill the fields for this field.

pods_field_display() will show a formatted link.