Switch tests to using wpSetupBeforeClass and wpTearDownAfterClass
WordPress apparently uses SetupBeforeClass/TearDownAfterClass and does some other funny stuff in the test suite.  For some reason the tests worked fine on everything but PHP 5.3, it seems like
it should have the same results on everything but I'm not going to spend the time tracking down an issue that will probably end up being a WordPress problem that isn't affecting anything else.