Fixed the CLI export/export-item commands.
When the export_data method of a pod without find parameters is called, and no 'params' are passed explicitly to it, an empty array is returned. Assuming this is the intended behavior, I added find parameters to the pod created by the export CLI command.

The export_data_to_file method did not set the 'columns' field of the data, which is required for any of the exports to work, so I added code to derive the columns from the supplied data.

The export_data_to_file method always forwarded the data enclosed in an array. This caused the JSON export-item to enclose the item in an array, the XML export to report only 1 item, and the CSV export not to work at all. Forwarding the data leads to a more expected result for the JSON export-item, and required handling the 'single' parameter for the CSV/XML export-item to continue to work.