pods->save() does not clear fields unless array is provided
`pods("pod")->save("field", null);` will never fire. In the same way `pods("pod")->save("field.id", null)` will never clear a relationship due to the current `$value==null` check. It will only work if it's called like `pods("pod")->save(array("field" => null));` which is rather confusing behaviour.

The current `$value != null` check doesn't seem to make sense, and really is supposed to check for `$data` being a string, and if so then transforming it to an array, right?

This proposed change checks instead if `$data` is an array, and if not, changes `$data` to an array, no matter the value of `$value`. This allows for clearing fields with the shorthand syntax.