Allow Auto Template cache to be refreshed on save
Using the 'update_option' action and checking for a specific option to be set has too many assumptions and wasn't working on my test.  This switches to the 'pods_cache_flushed' action.

Additionally, if the reset attempt did ever get made the actual transient name was wrong.  The additional transients that get cleared probably aren't necessary anymore but do not hurt anything.